requirements.txt
fastapi[standard]
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
bcrypt
pydantic
python-decouple
sqlmodel
passlib
fastapi_mail
python-socketio
python-jose
pydantic[email]
psycopg2
google
faker
aiofiles
sqlalchemy-continuum
SQLAlchemy-Utils
